+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 ## [0.0.1] - First Release.
+## [0.1.0] - Updated properties and button drawing method  
+
+Added new properties for foreground color, background color, heroTag, highlightElevation, mini (miniMode), and tooltip. Added more documentation (again based on the original Floating Action Button code).